Autotest/autotest/requirements.txt
django
djangorestframework
django-cors-headers
xlrd
xlwt
psycopg2-binary
pyyaml
jinja2
